+++ tests/dummy/app/models/ember-flexberry-dummy-comment-vote.js
@@ -1,15 +1,14 @@
 import DS from 'ember-data';
-import BaseModel from 'ember-flexberry-data/models/model';
-import Proj from 'ember-flexberry-data';
+import { Projection } from 'ember-flexberry-data';
 
-var Model = BaseModel.extend({
+var Model = Projection.Model.extend({
   // Inversed relationship for ember-flexberry-dummy-comment.userVotes.
   // It's not a property for flexberry-lookup component.
   comment: DS.belongsTo('ember-flexberry-dummy-comment', {
     inverse: 'userVotes',
     async: false
   }),
-  //voteType: DS.attr('ember-flexberry-dummy-vote-type'),
+  voteType: DS.attr('ember-flexberry-dummy-vote-type'),
 
   // This property is for flexberry-lookup component. No inverse relationship here.
   applicationUser: DS.belongsTo('ember-flexberry-dummy-application-user', {
@@ -24,9 +23,9 @@ var Model = BaseModel.extend({
 
 // Edit form projection.
 Model.defineProjection('CommentVoteE', 'ember-flexberry-dummy-comment-vote', {
-  voteType: Proj.attr('Vote type'),
-  applicationUser: Proj.belongsTo('ember-flexberry-dummy-application-user', 'Application user', {
-    name: Proj.attr('Name', {
+  voteType: Projection.attr('Vote type'),
+  applicationUser: Projection.belongsTo('ember-flexberry-dummy-application-user', 'Application user', {
+    name: Projection.attr('Name', {
       hidden: true
     })
   }, {