+++ tests/dummy/app/models/ember-flexberry-dummy-suggestion-type.js
@@ -1,8 +1,7 @@
 import DS from 'ember-data';
-import BaseModel from 'ember-flexberry-data/models/model';
-import Proj from 'ember-flexberry-data';
+import { Projection } from 'ember-flexberry-data';
 
-var Model = BaseModel.extend({
+var Model = Projection.Model.extend({
   name: DS.attr('string'),
   moderated: DS.attr('boolean'),
 
@@ -31,19 +30,19 @@ var Model = BaseModel.extend({
 
 // Edit form projection.
 Model.defineProjection('SuggestionTypeE', 'ember-flexberry-dummy-suggestion-type', {
-  name: Proj.attr('Name'),
-  moderated: Proj.attr('Moderated'),
-  parent: Proj.belongsTo('ember-flexberry-dummy-suggestion-type', 'Parent', {
-    name: Proj.attr('Name', {
+  name: Projection.attr('Name'),
+  moderated: Projection.attr('Moderated'),
+  parent: Projection.belongsTo('ember-flexberry-dummy-suggestion-type', 'Parent', {
+    name: Projection.attr('Name', {
       hidden: true
     })
   }, {
     displayMemberPath: 'name'
   }),
-  localizedTypes: Proj.hasMany('ember-flexberry-dummy-localized-suggestion-type', 'Localized types', {
-    name: Proj.attr('Name'),
-    localization: Proj.belongsTo('ember-flexberry-dummy-localization', 'Localization', {
-      name: Proj.attr('Name', {
+  localizedTypes: Projection.hasMany('ember-flexberry-dummy-localized-suggestion-type', 'Localized types', {
+    name: Projection.attr('Name'),
+    localization: Projection.belongsTo('ember-flexberry-dummy-localization', 'Localization', {
+      name: Projection.attr('Name', {
         hidden: true
       })
     }, {
@@ -54,10 +53,10 @@ Model.defineProjection('SuggestionTypeE', 'ember-flexberry-dummy-suggestion-type
 
 // List form projection.
 Model.defineProjection('SuggestionTypeL', 'ember-flexberry-dummy-suggestion-type', {
-  name: Proj.attr('Name'),
-  moderated: Proj.attr('Moderated'),
-  parent: Proj.belongsTo('ember-flexberry-dummy-suggestion-type', 'Parent', {
-    name: Proj.attr('Name', {
+  name: Projection.attr('Name'),
+  moderated: Projection.attr('Moderated'),
+  parent: Projection.belongsTo('ember-flexberry-dummy-suggestion-type', 'Parent', {
+    name: Projection.attr('Name', {
       hidden: true
     })
   }, {
@@ -67,26 +66,26 @@ Model.defineProjection('SuggestionTypeL', 'ember-flexberry-dummy-suggestion-type
 
 // Projection for lookup example on window customization.
 Model.defineProjection('SettingLookupExampleView', 'ember-flexberry-dummy-suggestion-type', {
-  name: Proj.attr('Name'),
-  moderated: Proj.attr('Moderated')
+  name: Projection.attr('Name'),
+  moderated: Projection.attr('Moderated')
 });
 
 // Projection for lookup example on window customization.
 Model.defineProjection('CustomizeLookupWindowExampleView', 'ember-flexberry-dummy-suggestion-type', {
-  name: Proj.attr('Name'),
-  moderated: Proj.attr('Moderated')
+  name: Projection.attr('Name'),
+  moderated: Projection.attr('Moderated')
 });
 
 // Projection for lookup with limit function example.
 Model.defineProjection('LookupWithLimitFunctionExampleView', 'ember-flexberry-dummy-suggestion-type', {
-  name: Proj.attr('Name'),
-  moderated: Proj.attr('Moderated')
+  name: Projection.attr('Name'),
+  moderated: Projection.attr('Moderated')
 });
 
 // Projection for lookup in dropdown mode example.
 Model.defineProjection('DropDownLookupExampleView', 'ember-flexberry-dummy-suggestion-type', {
-  name: Proj.attr('Name'),
-  moderated: Proj.attr('Moderated')
+  name: Projection.attr('Name'),
+  moderated: Projection.attr('Moderated')
 });
 
 export default Model;