+++ CHANGELOG.md
@@ -1,15 +1,33 @@
 # Change Log
 
 ## [Unreleased]
+
+## [0.3.0] - 19.07.2016
 ### Added
 - Support `polymorphic` options for `belongsTo` relationships.
 - Support special 'id' attribute at query language:
 
-  ```javascript 
+  ```javascript
   new SimplePredicate('manager', 'eq', key) === new SimplePredicate('manager.id', 'eq', key)
   ```
+- Classes to support work in offline mode. See `Offline` namespace in [api documentation](http://flexberry.github.io/Documentation/master/modules/ember-flexberry-data.html).
+
+### Changed
+- The addon is no longer exports `Projection` namespace with following extra changes:
+    - `Projection`, `Offline`, `Adapter`, `Serializer`, `Query`, `Utils` namespaces are exporting now.
+    - `Projection.Store` mixin renamed to `Projection.StoreMixin`.
+    - `Projection.Adapter` mixin renamed to `Projection.AdapterMixin`.
+    - All adapters were included into `Adapter` napespace.
+    - All serializers were included into `Serializer` napespace.
+    - All client query language classes were included into `Query` napespace.
+    - `Information` class was included into `Utils` napespace.
+  For more information please see [api documentation](http://flexberry.github.io/Documentation/master/modules/ember-flexberry-data.html).
+
 ### Fixed
 - Using `QueryBuilder.byId` method for OData query language adapter.
+
+### Known issues
+- Using properties for relationships in projections may cause errors when apply `selectByProjection` method of query language for reading models.
 
 ## [0.2.0] - 01.07.2016
 ### Added
@@ -38,4 +56,4 @@ The first release of ember-flexberry-data!
 ### Added
 - Implementation of _projections_ - predefined named set of model attributes.
 - Implementation of abstract query language and adapters for OData and JSON.
-- Implementations of `DS.Store` and `DS.Adapter` with support of _projections_. 
+- Implementations of `DS.Store` and `DS.Adapter` with support of _projections_.